refactor: [Task 1.3] Extract CompactDisplayConfig to dedicated module
